edited parts in project

+++ Galaxy/src/app/components/admin/category/main-category-admin/main-category-admin.component.ts
@@ -1,5 +1,8 @@
 import { Component } from '@angular/core';
 import { UnitService } from '../../../../services/unit.service';
+import { ConfirmMessageComponent } from '../../../shared-componentes/confirm-message/confirm-message.component';
+import { MatDialog } from '@angular/material/dialog';
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
   selector: 'app-main-category-admin',
@@ -7,7 +10,9 @@ import { UnitService } from '../../../../services/unit.service';
   styleUrl: './main-category-admin.component.css'
 })
 export class MainCategoryAdminComponent {
-  constructor(private unit: UnitService) { }
+  constructor(private unit: UnitService, private route: ActivatedRoute,
+    private router: Router,
+    public dialog: MatDialog) { }
   categories: any;
 
 
@@ -16,18 +21,37 @@ export class MainCategoryAdminComponent {
   }
 
   Deletecategory(category:any){
-    this.unit.category.deleteCategory(category.id).subscribe(()=>{
-      category.isDeleted = true;
-    },error=>{
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to delete ${category.name}?`,
+      title: 'Delete category'  }
+    });
 
-    })
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.unit.category.deleteCategory(category.id).subscribe(()=>{
+          category.isDeleted = true;
+        },error=>{
+
+        })
+      }
+    });
   }
   RetrieveCategory(category: any) {
-    this.unit.category.retrieveCategory(category.id).subscribe(() => {
-      category.isDeleted = false;
-    }, error => {
-      console.error('Error retrieving brand', error);
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to retrive ${category.name}?`,
+      title: 'retrive category'  }
     });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.unit.category.retrieveCategory(category.id).subscribe(() => {
+          category.isDeleted = false;
+        }, error => {
+          console.error('Error retrieving brand', error);
+        });
+      }
+    });
+
   }
 
   GetCategories(): void {
