+++ Galaxy/src/app/components/admin/subcategory/main-admin-subcategory/main-admin-subcategory.component.ts
@@ -1,13 +1,16 @@
 import { Component } from '@angular/core';
 import { UnitService } from '../../../../services/unit.service';
-
+import { ConfirmMessageComponent } from '../../../shared-componentes/confirm-message/confirm-message.component';
+import { MatDialog } from '@angular/material/dialog';
+import { ActivatedRoute, Router } from '@angular/router';
 @Component({
   selector: 'app-main-admin-subcategory',
   templateUrl: './main-admin-subcategory.component.html',
   styleUrl: './main-admin-subcategory.component.css'
 })
 export class MainAdminSubcategoryComponent {
-  constructor(private unit: UnitService) { }
+  constructor(private unit: UnitService ,  private router: Router,
+    public dialog: MatDialog) { }
   subcategories: any;
 
 
@@ -16,14 +19,35 @@ export class MainAdminSubcategoryComponent {
   }
 
   DeleteSubcategory(subcategory:any){
-    this.unit.subcategory.DeleteSubCategory(subcategory.id).subscribe(()=>{
-      subcategory.isDeleted = true;
-    })
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to delete ${subcategory.name}?`,
+      title: 'Delete SubCategory'  }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+
+      if (result) {
+        this.unit.subcategory.DeleteSubCategory(subcategory.id).subscribe(()=>{
+          subcategory.isDeleted = true;
+        })
+      }
+    });
+
   }
   RetriveSubCategory(subcategory:any){
-    this.unit.subcategory.RetriveSubCategory(subcategory.id).subscribe(()=>{
-      subcategory.isDeleted = false;
-    })
+    const dialogRef = this.dialog.open(ConfirmMessageComponent, {
+      data: { message: `Are you sure you want to retrive ${subcategory.name}?`,
+      title: 'retrive SubCategory'  }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.unit.subcategory.RetriveSubCategory(subcategory.id).subscribe(()=>{
+          subcategory.isDeleted = false;
+        })
+      }
+    });
+
   }
   GetSubCategory():void{
     this.subcategories=this.unit.subcategory.GetAdminSubCategories().subscribe((subcategory: any) => {
@@ -42,6 +66,6 @@ export class MainAdminSubcategoryComponent {
       console.error('Error updating subcategory:', error);
     });
   }
- 
+
 
 }
